+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/compute/GetPolicyComplianceForJob.java
@@ -1,0 +1,43 @@
+// Code generated from OpenAPI specs by Databricks SDK Generator. DO NOT EDIT.
+
+package com.databricks.sdk.service.compute;
+
+import com.databricks.sdk.support.Generated;
+import com.databricks.sdk.support.QueryParam;
+import com.databricks.sdk.support.ToStringer;
+import java.util.Objects;
+
+/** Get job policy compliance */
+@Generated
+public class GetPolicyComplianceForJob {
+  /** The ID of the job whose compliance status you are requesting. */
+  @QueryParam("job_id")
+  private Long jobId;
+
+  public GetPolicyComplianceForJob setJobId(Long jobId) {
+    this.jobId = jobId;
+    return this;
+  }
+
+  public Long getJobId() {
+    return jobId;
+  }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null || getClass() != o.getClass()) return false;
+    GetPolicyComplianceForJob that = (GetPolicyComplianceForJob) o;
+    return Objects.equals(jobId, that.jobId);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(jobId);
+  }
+
+  @Override
+  public String toString() {
+    return new ToStringer(GetPolicyComplianceForJob.class).add("jobId", jobId).toString();
+  }
+}