+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/compute/ListPolicyFamilies.java
@@ -9,8 +9,8 @@ import java.util.Objects;
 
 /** List policy families */
 @Generated
-public class ListPolicyFamiliesRequest {
-  /** The max number of policy families to return. */
+public class ListPolicyFamilies {
+  /** Maximum number of policy families to return. */
   @QueryParam("max_results")
   private Long maxResults;
 
@@ -18,7 +18,7 @@ public class ListPolicyFamiliesRequest {
   @QueryParam("page_token")
   private String pageToken;
 
-  public ListPolicyFamiliesRequest setMaxResults(Long maxResults) {
+  public ListPolicyFamilies setMaxResults(Long maxResults) {
     this.maxResults = maxResults;
     return this;
   }
@@ -27,7 +27,7 @@ public class ListPolicyFamiliesRequest {
     return maxResults;
   }
 
-  public ListPolicyFamiliesRequest setPageToken(String pageToken) {
+  public ListPolicyFamilies setPageToken(String pageToken) {
     this.pageToken = pageToken;
     return this;
   }
@@ -40,7 +40,7 @@ public class ListPolicyFamiliesRequest {
   public boolean equals(Object o) {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
-    ListPolicyFamiliesRequest that = (ListPolicyFamiliesRequest) o;
+    ListPolicyFamilies that = (ListPolicyFamilies) o;
     return Objects.equals(maxResults, that.maxResults) && Objects.equals(pageToken, that.pageToken);
   }
 
@@ -51,7 +51,7 @@ public class ListPolicyFamiliesRequest {
 
   @Override
   public String toString() {
-    return new ToStringer(ListPolicyFamiliesRequest.class)
+    return new ToStringer(ListPolicyFamilies.class)
         .add("maxResults", maxResults)
         .add("pageToken", pageToken)
         .toString();