+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/compute/PolicyFamiliesImpl.java
@@ -24,7 +24,7 @@ class PolicyFamiliesImpl implements PolicyFamiliesService {
   }
 
   @Override
-  public ListPolicyFamiliesResponse list(ListPolicyFamiliesRequest request) {
+  public ListPolicyFamiliesResponse list(ListPolicyFamilies request) {
     String path = "/api/2.0/policy-families";
     Map<String, String> headers = new HashMap<>();
     headers.put("Accept", "application/json");