+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/compute/PolicySortColumn.java
@@ -5,7 +5,7 @@ package com.databricks.sdk.service.compute;
 import com.databricks.sdk.support.Generated;
 
 @Generated
-public enum ListSortColumn {
+public enum PolicySortColumn {
   POLICY_CREATION_TIME,
   POLICY_NAME,
 }