+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/dashboards/TextAttachment.java
@@ -13,6 +13,10 @@ public class TextAttachment {
   @JsonProperty("content")
   private String content;
 
+  /** */
+  @JsonProperty("id")
+  private String id;
+
   public TextAttachment setContent(String content) {
     this.content = content;
     return this;
@@ -22,21 +26,30 @@ public class TextAttachment {
     return content;
   }
 
+  public TextAttachment setId(String id) {
+    this.id = id;
+    return this;
+  }
+
+  public String getId() {
+    return id;
+  }
+
   @Override
   public boolean equals(Object o) {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
     TextAttachment that = (TextAttachment) o;
-    return Objects.equals(content, that.content);
+    return Objects.equals(content, that.content) && Objects.equals(id, that.id);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(content);
+    return Objects.hash(content, id);
   }
 
   @Override
   public String toString() {
-    return new ToStringer(TextAttachment.class).add("content", content).toString();
+    return new ToStringer(TextAttachment.class).add("content", content).add("id", id).toString();
   }
 }